+++ docs/CHANGELOG.rst
@@ -34,6 +34,7 @@ Changed
 * Move testing utilities from ``resolwe_bio.tests.processes.utils`` to
   ``resolwe_bio.utils.test``
 * Update Cuffdiff processor inputs and results table parsing
+* Refactor processes to use the updated ``resolwe.flow.executors.run`` command
 
 Fixed
 -----